move hypervisor/utils.go to lib/utils/
